git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@27384 aa0bda67-78ed-0310-a382-c5338b0f89ba
Former-commit-id: b6d12a408f6d0bc777c581be4f1b99a27c4678c5

+++ common/core/src/de/uniol/inf/is/odysseus/core/metadata/TimeInterval.java
@@ -112,9 +112,9 @@ final public class TimeInterval extends AbstractBaseMetaAttribute implements
 	public <K> K getValue(int subtype, int index) {
 		switch (index) {
 		case 0:
-			return (K) (Long) (start.isInfinite()?null:start.point);
+			return (K) (start.isInfinite()?null:start.point);
 		case 1:
-			return (K) (Long) (end.isInfinite()?null:end.point);
+			return (K) (end.isInfinite()?null:end.point);
 		}
 		return null;
 	}
