+++ server/imagejcv/src/de/uniol/inf/is/odysseus/imagejcv/functions/TestImageFunction.java
@@ -66,7 +66,7 @@ public class TestImageFunction extends AbstractFunction<ImageJCV> {
 		{
 			int newIndex = y * newImage.getWidthStep() + x * newImage.getNumChannels() * 2;
 			
-			short val = (short)(cur++);
+			short val = (cur++);
 			if (flip)
 				val ^= 0xFFFF;
